RequestDTO에서 ReponseDTO로 수정
