+++ src/main/java/com/example/jobala/comp/CompRepository.java
@@ -4,6 +4,7 @@ import com.example.jobala.jobopen.Jobopen;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.Query;
 import lombok.RequiredArgsConstructor;
+import org.qlrm.mapper.JpaResultMapper;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,12 +15,18 @@ import java.util.List;
 public class CompRepository {
     private final EntityManager em;
 
-    public CompRequest.scoutListDTO scoutList(){
+    public CompResponse.scoutListDTO scoutList(){
         String q = """
                 SELECT ut.name, rt.resume_title, ut.age, ut.address, rt.career
                 FROM resume_tb rt
                 INNER JOIN user_tb ut ON rt.user_id = ut.id
                 """;
+
+        Query query = em.createNativeQuery(q);
+        JpaResultMapper jpaResultMapper = new JpaResultMapper();
+
+
+
         return null;
     }
