+++ src/main/java/com/example/jobala/comp/CompResponse.java
@@ -1,5 +1,19 @@
 package com.example.jobala.comp;
 
 
+import lombok.Data;
+
+import java.sql.Date;
+
 public class CompResponse {
+
+    @Data
+    public static class scoutListDTO{
+        private String name;
+        private String resumeTitle;
+        private Date age;
+        private String address;
+        private String career;
+    }
+
 }
